\name{tmCaption}
\alias{tmCaption}
\title{
  Gets the attribute 'caption' from a 'TableMonster' class object
}
\description{
  Gets the attribute 'caption' from a 'TableMonster' class object
}
\usage{
  tmCaption(x)
}
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\value{
  A character string
}
\author{Grant Izmirlian}

