\name{tmCaption<-}
\alias{tmCaption<-}
\title{
  Assignment function for the 'caption' attribute 
}
\description{
  Assignment function for the 'caption' attribute of an object of
  class 'TableMonster'
}
\usage{
  tmCaption(x) <- value
}
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
  \item{value}{
    A character string
  }
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\author{Grant Izmirlian}

