\name{tmCtypes}
\alias{tmCtypes}
\title{
  Gets the attribute 'ctypes' from a 'TableMonster' class object
}
\description{
  Gets the attribute 'ctypes' from a 'TableMonster' class object
}
\usage{
  tmCtypes(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\value{
  A character vector the same length as the number of columns of the
  table, having entries "n" or "c", meaning "numeric" or "character"
}
\author{Grant Izmirlian}

