\name{tmDigits}
\alias{tmDigits}
\title{
  Gets the attribute 'digits' from a 'TableMonster' class object
}
\description{
  Gets the attribute 'digits' from a 'TableMonster' class object
}
\usage{
  tmDigits(x)
}
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\value{
  A numeric vector of length equal to the number of columns in the table
}
\author{Grant Izmirlian}

