\name{tmHeadings}
\alias{tmHeadings}
\title{
  Gets the attribute 'headings' from a 'TableMonster' class object
}
\description{
  Gets the attribute 'headings' from a 'TableMonster' class object
}
\usage{
tmHeadings(x)
}
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
}
\value{
  The 'headings' attribute of a 'TableMonster' object, a vector of
  character strings of length equal to the number of columns of the table.
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\author{Grant Izmirlian}
