\name{tmHeadings<-}
\alias{tmHeadings<-}
\title{
  Assignment function for the 'headings' attribute
}
\description{
  Assignment function for the 'headings' attribute for an object of
  class 'TableMonster'
}
\usage{
  tmHeadings(x) <- value
}
\arguments{
  \item{x}{
    An object of class 'TableMonster'
  }
  \item{value}{
    A vector of character strings of length equal to the number of
    columns in the table
  }
}
\details{
  This is a required attribute for an object of class 'TableMonster'
}
\author{Grant Izmirlian}
\references{
  <https://www.youtube.com/watch?v=CM1TaNVnh58>
}
